fix :add account deletion config in the profile API
Account deletion should be configurable based on the
`ENABLE_ACCOUNT_DELETION` feature flag.

To enable or disable the account deletion feature in the
[frontend-app-account](https://github.com/openedx/frontend-app-account),
it is essential to make the configuration available through the API.

Including the `enable_account_deletion` flag in the existing User
Profile API is logical as it aligns with the user profile's scope. Since
other user-related configurations are already managed through the API,
it is the correct choice to handle account deletion settings.